requirements.txt
sentence-transformers
qdrant-client
duckdb
neo4j
pandas
rich
python-dotenv
openai